relocate CompositeCrcFileChecksum and related classes to avoid classpath conflict
