+++ hadoop-ozone/common/src/main/java/org/apache/hadoop/ozone/client/checksum/CompositeCrcFileChecksum.java
@@ -15,20 +15,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.fs;
+package org.apache.hadoop.ozone.client.checksum;
 
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
 
-import org.apache.hadoop.classification.InterfaceAudience;
-import org.apache.hadoop.classification.InterfaceStability;
+import org.apache.hadoop.fs.FileChecksum;
 import org.apache.hadoop.fs.Options.ChecksumOpt;
-import org.apache.hadoop.util.CrcUtil;
+import org.apache.hadoop.hdds.annotation.InterfaceAudience;
+import org.apache.hadoop.hdds.annotation.InterfaceStability;
 import org.apache.hadoop.util.DataChecksum;
 
 /** Composite CRC. */
-@InterfaceAudience.LimitedPrivate({"HDFS"})
+@InterfaceAudience.Private
 @InterfaceStability.Unstable
 public class CompositeCrcFileChecksum extends FileChecksum {
   public static final int LENGTH = Integer.SIZE / Byte.SIZE;
